Decompose Settings.tsx into tab files + shared controls
Extract Toggle, Select, NumberInput, TextInput, Field into FormControls.tsx.
Extract General, Display, Advanced, ApiKeys, Cookies, TokenAccounts, About
tabs into surfaces/settings/tabs/. Settings.tsx drops from 1664 to ~400 lines.
ProvidersTab and its coupled helpers remain in Settings.tsx.
