+++ apps/desktop-tauri/src/components/FormControls.tsx
@@ -1,0 +1,129 @@
+import type React from "react";
+
+// ── tiny reusable controls ──────────────────────────────────────────
+
+export function Toggle({
+  checked,
+  onChange,
+  disabled,
+}: {
+  checked: boolean;
+  onChange: (v: boolean) => void;
+  disabled?: boolean;
+}) {
+  return (
+    <button
+      role="switch"
+      aria-checked={checked}
+      className={`toggle ${checked ? "toggle--on" : ""}`}
+      disabled={disabled}
+      onClick={() => onChange(!checked)}
+    />
+  );
+}
+
+export function Select({
+  value,
+  options,
+  onChange,
+  disabled,
+}: {
+  value: string;
+  options: { value: string; label: string }[];
+  onChange: (v: string) => void;
+  disabled?: boolean;
+}) {
+  return (
+    <select
+      className="select"
+      value={value}
+      disabled={disabled}
+      onChange={(e) => onChange(e.target.value)}
+    >
+      {options.map((o) => (
+        <option key={o.value} value={o.value}>
+          {o.label}
+        </option>
+      ))}
+    </select>
+  );
+}
+
+export function NumberInput({
+  value,
+  min,
+  max,
+  step,
+  onChange,
+  disabled,
+}: {
+  value: number;
+  min?: number;
+  max?: number;
+  step?: number;
+  onChange: (v: number) => void;
+  disabled?: boolean;
+}) {
+  return (
+    <input
+      type="number"
+      className="number-input"
+      value={value}
+      min={min}
+      max={max}
+      step={step}
+      disabled={disabled}
+      onChange={(e) => {
+        const n = Number(e.target.value);
+        if (!Number.isNaN(n)) onChange(n);
+      }}
+    />
+  );
+}
+
+export function TextInput({
+  value,
+  placeholder,
+  onChange,
+  disabled,
+}: {
+  value: string;
+  placeholder?: string;
+  onChange: (v: string) => void;
+  disabled?: boolean;
+}) {
+  return (
+    <input
+      type="text"
+      className="text-input"
+      value={value}
+      placeholder={placeholder}
+      disabled={disabled}
+      onChange={(e) => onChange(e.target.value)}
+    />
+  );
+}
+
+// ── field row ────────────────────────────────────────────────────────
+
+export function Field({
+  label,
+  description,
+  children,
+}: {
+  label: string;
+  description?: string;
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="settings-field">
+      <div className="settings-field__text">
+        <span className="settings-field__label">{label}</span>
+        {description && (
+          <span className="settings-field__desc">{description}</span>
+        )}
+      </div>
+      <div className="settings-field__control">{children}</div>
+    </div>
+  );
+}
